# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Farm Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notion/list-databases:
    get:
      operationId: listDatabases
      tags: [notion]
      summary: List all Notion databases in workspace
      description: Returns all databases the user has shared with the Notion integration
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of databases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Notion not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notion/inspect-database:
    get:
      operationId: inspectDatabase
      tags: [notion]
      summary: Inspect Notion database properties
      description: Returns all properties (columns) from a Notion database with their types
      security:
        - clerkAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [panen, expenses]
        - name: databaseId
          in: query
          required: false
          schema:
            type: string
          description: Notion database ID to inspect. If provided, used directly. Otherwise tries saved mapping then name search.
      responses:
        "200":
          description: Database properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectDatabaseResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Notion not connected or database not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notion/harvest-dropdown-options:
    get:
      operationId: getHarvestDropdownOptions
      tags: [notion]
      summary: Get dropdown options for harvest form
      description: Returns list of Pindah Tanam pages from Notion for the Area field
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Harvest dropdown options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HarvestDropdownOptions"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Notion not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notion/add-harvest:
    post:
      operationId: addHarvest
      tags: [notion]
      summary: Add a new harvest entry to Notion Panen database
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddHarvestBody"
      responses:
        "201":
          description: Harvest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddHarvestResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Notion not connected or Panen database not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    NotionOption:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
...
    DatabaseProperty:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        relatedDatabaseId:
          type: ["string", "null"]
      required:
        - id
        - name
        - type
    InspectDatabaseResponse:
      type: object
      properties:
        databaseId:
          type: string
        databaseName:
          type: string
        properties:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseProperty"
      required:
        - databaseId
        - databaseName
        - properties
...
    DatabaseListItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        iconEmoji:
          type: ["string", "null"]
      required:
        - id
        - name
    DatabaseListResponse:
      type: object
      properties:
        databases:
          type: array
          items:
            $ref: "#/components/schemas/DatabaseListItem"
      required:
        - databases
...
    HarvestDropdownOptions:
      type: object
      properties:
        pindahTanam:
          type: array
          items:
            $ref: "#/components/schemas/NotionOption"
        labaRugi:
          type: array
          items:
            $ref: "#/components/schemas/NotionOption"
      required:
        - pindahTanam
        - labaRugi
    AddHarvestBody:
      type: object
      properties:
        kegiatan:
          type: string
          description: Harvest activity name (Title property)
        jumlahPanen:
          type: number
          description: Harvest quantity in kg
        hargaJualPerKg:
          type: number
          description: Selling price per kg
        kualitas:
          type: string
          description: Harvest quality grade
        channelPenjualan:
          type: string
          description: Sales channel
        pindahTanamId:
          type: string
          description: Page ID of the related Pindah Tanam entry (Area Pindah Tanam relation)
        labaRugiId:
          type: string
          description: Page ID of the related Laba Rugi entry (Area Laba Rugi relation)
      required:
        - kegiatan
        - jumlahPanen
        - hargaJualPerKg
        - kualitas
        - channelPenjualan
        - pindahTanamId
        - labaRugiId
    AddHarvestResult:
      type: object
      properties:
        success:
          type: boolean
        pageId:
          type: string
          description: Notion page ID of the newly created harvest entry
      required:
        - success
        - pageId
...